feat: wx cache

+++ triones-weixin-base/src/main/kotlin/com/trionesdev/weixin/base/model/AccessTokenResponse.kt
@@ -12,5 +12,5 @@ class AccessTokenResponse : BaseResponse() {
     var accessToken: String? = null
 
     @JsonProperty(value = "expires_in")
-    var expiresIn: Int? = null
+    var expiresIn: Long? = null
 }
